Draft/exciting worker

+++ pages/api/cancelapt.js
@@ -10,7 +10,7 @@ export default async function handler(req, res) {
     const auth = await verifyJWT(parsed.token, req.body.adminID);
     if (auth.auth && auth.act === 'admin' && auth.id === adminID) {
         const apt = await model.Appointments.findOne({where: { id: aptID }})
-        if(apt.adminID === adminID) {
+        if(apt && apt.adminID === adminID) {
             const deleted = await model.Appointments.destroy({ where: { id: aptID }})
             if(deleted > 0) {
                 res.status(200).send()
